modal/refactoring
Я закінчив рефакторінг. Файл modal.js видалив. Перевір свою функцію вона у файлі add-muvie-to-database папка service.
