+++ src/index.js
@@ -15,13 +15,11 @@ const global = {
 
 //Віше определение всех глобал переменніх проекта
 
-// import './js/modal';
 import './js/service/init-firebase';
 
 // import './js/goToMyLibrary';
 import './js/pagination';
 import './js/header-library';
-import './js/modal';
 import './js/modal_tm';
