+++ src/js/service/backdrop-modal-close.js
@@ -1,0 +1,7 @@
+import { closeModal } from "./close-modal";
+
+export function backdropClick(event) {
+  if (event.currentTarget === event.target) {
+    closeModal();
+  }
+}
